requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
transformers
accelerate
imageio-ffmpeg
gradio==4.44.1
pyannote.audio
